refactor attention, improve explicit flash attention performance with the prestore_kv_flag
