+++ yalis/attention/backends.py
@@ -1,0 +1,4 @@
+# These imports trigger @register_attention decorators
+from . import sdpa
+from . import flash
+
