+++ yalis/model.py
@@ -16,7 +16,8 @@ def get_model(
     device="cuda",
     use_intra_head_parallelism=False,
     explicitly_use_flash_kernel=False,
-    use_paged_kv_caching=False
+    use_paged_kv_caching=False,
+    prestore_kv_cache=True
 ):
     tensor_parallel = dist.get_world_size() > 1
     if tensor_parallel and dist.get_rank() == 0:
@@ -32,6 +33,7 @@ def get_model(
     config.use_intra_head_parallelism = use_intra_head_parallelism
     config.explicitly_use_flash_kernel = explicitly_use_flash_kernel
     config.use_paged_kv_caching = use_paged_kv_caching
+    config.prestore_kv_cache = prestore_kv_cache
 
     with _EmptyInit(enabled=(not random_init)):
         model = GPT(config).to(model_dtype)
