+++ yalis/utils.py
@@ -3,6 +3,8 @@ import torch.distributed as dist
 
 
 def print_rank0(msg):
+    if not dist.is_initialized():
+        raise RuntimeError("Distributed process group is not initialized. Make sure init_distributed() is called before using this function.")
     if dist.get_rank() == 0:
         print(f"{msg}")
